backend/requirements.txt
fastapi
uvicorn
boto3
psycopg2-binary
python-dotenv
pandas
langchain
langchain-community
rank_bm25
google-genai
